Overrides OAuthRequest::getRequestBodyOfMultipart() in BP_OAuthServer
When the OAuth library creates a URL base from which to verify/create a
signature hash, it sometimes attempts to include the POST body. This causes
a discrepancy between signatures on the client side (when there is no POST
body) and the server side (where there is). So we return null instead.