Ensure that Infinity's activation hook is fired when activating cbox-theme
Infinity uses the infinity_dashboard_activated hook to run some setup
routines at activation. However, Infinity itself only fires the hook
when activating themes in the traditional way, from themes.php. When
activating the theme with switch_theme(), we thus must have a way of
manually firing the hook.

However, the hook only does any good when it's loaded alongside of
the theme files themselves, which means that we have to wait for
the subsequent pageload after the theme has been switched in the
options table.

The solution in this changeset is to set a _cbox_theme_activated flag
after using switch_theme(), and then to check for its presence when
looking at CBOX admin pages. If the flag is found, the hook is fired,
and the flag deleted.