Show an admin notice when both BP bundled forums and bbPress are enabled.
The bundled forums component in BuddyPress combined with bbPress 2 leads to
conflicts between the two plugins when both are enabled.

We show a notice if both BuddyPress' bundled forums and bbPress are enabled
so site admins are aware of the potential conflict with instructions of
what they can do to address the issue.

Note: Temporary documentation link in notice needs to be published.

See #36.