Fix fatal error.
On a fresh single CBOX install, BuddyPress is not activated yet.

BP has some abstraction code for multisite and CBOX was using
get_blog_option() in plugin-install.php before BP was activated.

This caused a fatal error.

To fix this, we do a check if multisite is enabled.  If so, then we use
get_blog_option().  If not, we're on a standard WP install, and we use
get_option().

Boo Ray.