Manipulate 'Updates' menu in WP Toolbar on frontend. Fixes #24.
CBOX doesn't load its plugin code on the frontend.  This can
interfere with the 'Updates' menu in the WP Toolbar as the update
count can be different than the update count in the admin area.

Hence, this commit removes the existing 'Updates' menu in the WP
Toolbar on the frontend.  It is replaced with our custom 'Updates'
menu, which removes the update counts.