Removes Bebop from list of Optional plugins
We still think that Bebop is great. But behind-the-scenes issues make it
unclear whether the maintainers of Bebop will be able to keep it up to date
with upcoming changes to APIs offered by Twitter and other providers. To
provide the best out-of-the-box experience for Commons In A Box users, we're
removing Bebop from the list of plugins that is managed by the software.

Existing installations that are using Bebop will continue to work as before,
though Commons In A Box will no longer manage updates and activation.