Change how cbox_is_admin() determines if we're in the admin area.
This commit removes the usage of wp_get_referer() in cbox_is_admin()
as changes in WP 3.6.1 broke the usage of this function.

Instead, we now check if WP_NETWORK_ADMIN is defined or not.  This is
more reliable to determine whether we're in the admin dashboard than checking if the referer came from within the admin area.

Fixes #85.