Introduce CBox_Bulk_Plugin_Upgrader_Skin::_is_static() method.
This method is used to determine if a class method is called from a
static or object context.

Use this method in the after_updater() method, which can be called in
both static or object contexts.

Fixes previous fatal error when the after_updater() method was called
statically and referencing the class object incorrectly.