Allow non-logged-in users to view a private group's homepage.
CBOX has an option to set a group's default tab to "Forum".  If this
setting is enabled, the "Forum" tab is used instead of the default group
tab.

However, if a user is logged out and attempts to visit a private group,
that user is unable to view the private group homepage due to CBOX
overriding the default group tab to "Forum" and BuddyPress' login
redirection kicking in.

This commit remedies this by not setting the group tab to "Forum" when a
user is logged out and is viewing a private group.