Set `network = false` flag for certain CBOX plugins.
Plugins set to only be activated on the main site are:
- bbPress
- BP Group Announcements
- BuddyPress Docs
- BuddyPress Docs Wiki add-on
- Custom Profile Filters for BuddyPress
- External Group Blogs
- Invite Anyone

Note: bbPress and BuddyPress Docs will require some extra code to be
written so their WP admin bar links are visible when not on the main site.
Will probably add this code into CBOX directly.

Plugin activation to support this will be added in subsequent commits.

See #117.