Packages: Refactor package API registration.
Due to the usage of late static bindings, we need to refactor parts of the
package API registration code so inherited class data is separate from the
parent class.  What this means is the $config static property is required
in all extended classes of CBox_Package.  We're also removing some static
properties that are now going to be housed inside the $config property.

Also, since package data can be accessed without fully initializing the
package (such as before the package ID is saved in the database), we might
need to initialize the property values.  This commit introduces the
CBox_Package::get_props() and CBox_Package::set_props() static methods to
help do that.

See #119.