Plugin Install: Ensure we add dependent plugins to the installation queue for plugins that haven't been installed yet.
We use the Plugin Dependencies library to find out if a plugin has
dependent plugins.  However, that library only works for plugins that are
already installed in the filesystem.

However, since CBOX can install plugins that do not exist in the filesystem,
we need to parse the dependencies of these plugins before we can install
them.  If there are dependent plugins, these plugins need to be added to
the installation queue.

This is what this commit does.