Frontend: Register toolbar menu items for certain plugins on sub-sites.
In #117, CBOX plugins can now be either network-activated or
site-activated.  If a plugin is site-activated, but has registered toolbar
menu items on the main site, these menus will also need to be shown when
on a sub-site.

This commit adds toolbar menu items for bbPress, BuddyPress Docs and
BP Event Organiser when on a sub-site if these plugins are marked as
site-only.

See #117.