Allow administrative functionality to occur on either the main site admin or in the Network Admin area.
In commit 16ac4db, the intention was for packages to be able to set where
the admin area should be displayed in the main site admin area or the
Network Admin area.

However, in multisite, this commit allows the CBOX admin menu to be
displayed in *both* the main site admin area and in the Network Admin area.
This is necessary because if you switch packages or if you initially set
CBOX up for the first time, the CBOX admin area will need to be accessible.
If one package chooses to have the admin area display in the main site
admin area, but another package chooses to display in the Network Admin
area, this would offer an inconsistent user experience.

Therefore, commit 16ac4db can probably be removed and we offer the
CBOX admin area in both the main site and in the Network Admin area.