Plugins: Ensure our "Group Forums" category is created for bbPress during activation.
Previously, we were checking if the 'bb-config-location' was empty to
determine if CBOX should create the "Group Forums" category.

The problem is BuddyPress filters this option:
https://buddypress.trac.wordpress.org/browser/tags/2.9.4/src/bp-core/bp-core-options.php?marks=35#L34

So this option isn't empty.  To address this, we switch to using
file_exists() to see if the legacy bbPress bb-config.php file actually
exists or not.

See #133.