Fix BuddyPress component defaults after activation.
In commit a8b0a82, we introduced some functionality so packages can run
custom code when a plugin is activated.  This code runs on the
'activated_plugin' hook and only once.

However, BuddyPress has its own updater routine that runs on 'admin_init'.
Due to a multitude of factors that I won't explain, we need to run
the update routine for BuddyPress before BuddyPress can run theirs on the
next pageload.

This commit does this.

See #133.