Register email taxonomy during BuddyPress activation.
In commit 8f6b225, we moved to manually running the BuddyPress updater
routine.  However, this broke how BuddyPress creates email posts for
HTML emails.

To address this, we have to manually register BP's email taxonomy to ensure
BuddyPress email posts are attached to email situations.  See commit
43e1fa2 for a similar fix for the BP Group Email Subscription plugin.
Since we are now registering the email taxonomy when BuddyPress is
activated, we no longer need to register it during GES activation.

See #133.