Only exclude plugins if the CBOX plugin's 'hide' parameter is set to true.
Previously, we automatically displayed a plugin in the list table if on a
sub-site and if the plugin was not registered as a 'network' plugin.

With the new 'hide' parameter introduced in commit 1733cc4, we can exclude
plugins on a case-by-case basis and with greater granularity.

See https://github.com/cuny-academic-commons/openlab-theme/issues/183