Toolbar: Add fallback notification for bbPress on sub-sites.
In commit 9858001, we added support for formatting notifications if bbPress
is only activated on the main site and not network-wide.  This technique
relies on recording some forum post information into BP's notification
meta table.

If this information is not available, this commit adds a fallback link to
the logged-in user's notifications page, filtered by the 'bbp_new_reply'
notification type.

Also, adds compatibility for bbPress 2.6.0:
https://bbpress.trac.wordpress.org/changeset/6845/trunk/src/includes/extend/buddypress/notifications.php

See #117.