Plugins: Allow auto-activation for install-only plugins if the plugin is also a dependent.
Install-only plugins are generally not auto-activated after installation,
but if such a plugin is also a dependent, it needs to be activated.

For example, the OpenLab package uses the Braille plugin as a dependent
on the BP Braille plugin, but also as an install-only plugin for member
sites.

Commit a8e1c1a broke how the Braille plugin would not be auto-activated
when the BP Braille plugin is enabled.

This should now be addressed in this commit.

See #151.