bbPress: Allow logged-in users to bypass link limit.
The link limit was introduced in bbPress 2.6 and piggybacks off of the
blog comment moderation settings value, which defaults to only 2.

This means that if a user puts two links or more in their forum post,
their post would be automatically marked as 'pending' and placed into
the moderation queue.

See https://bbpress.trac.wordpress.org/ticket/3352

To workaround the blog comment's link limit, we will bypass the link
limit for forum posts entirely.

See #199.