bbPress: Add a template notice when a forum post is auto-moderated.
When a forum post contains a word that matches one in the comment
moderation list, the forum post is automatically marked as 'pending'
and requires moderation. However, the user is not informed as to why
the post failed to post. It just looks like the page refreshed.

The notice does not show up because bbPress does not retain notices
when a page is redirected.

See https://bbpress.trac.wordpress.org/ticket/3353.

This commit adds a template notice using BuddyPress's API since
their API can handle page redirections due to the use of cookies.

See #199.